fix chat_js done
