+++ app/javascript/packs/application.js
@@ -9,3 +9,4 @@ Turbolinks.start()
 ActiveStorage.start()
 
 import "stylesheets/application"
+import "controllers"
