+++ app/views/rooms/index.html.erb
@@ -1,4 +1,5 @@
-<div class="chat_room text-3xl">
+
+<div class="text-3xl" data-controller="chat" >
   Hello,<%= current_user.name %>
 </div>
 <div class="chat_box w-full flex ">
